requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
cachetools==5.2.0
celery==5.2.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
frozenlist==1.3.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
kombu==5.2.4
lxml==4.8.0
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
prompt-toolkit==3.0.30
protobuf==4.21.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
yarl==1.7.2
